Make integration tests work with factored-out TestWrapper
